Finalizing WebSocket server
We're reaching the end of the WebSocket server implementation, so we just need to complete the client and this project should be mostly finished. We may restructure it to support multiple different games via having "problem_words" file for each game we wish to support and, potentially, a dictionary or similar to convert phrases to their in-game names so mods could more easily be created.